+++ packages/gatsby/cache-dir/navigation.js
@@ -10,13 +10,23 @@ import { globalHistory } from "@reach/router/lib/history"
 import { parsePath } from "gatsby-link"
 
 // Convert to a map for faster lookup in maybeRedirect()
-const redirectMap = redirects.reduce((map, redirect) => {
-  map[redirect.fromPath] = redirect
-  return map
-}, {})
+
+const redirectMap = new Map()
+const redirectIgnoreCaseMap = new Map()
+
+redirects.forEach(redirect => {
+  if (redirect.ignoreCase) {
+    redirectIgnoreCaseMap.set(redirect.fromPath, redirect)
+  } else {
+    redirectMap.set(redirect.fromPath, redirect)
+  }
+})
 
 function maybeRedirect(pathname) {
-  const redirect = redirectMap[pathname]
+  let redirect = redirectMap.get(pathname)
+  if (!redirect) {
+    redirect = redirectIgnoreCaseMap.get(pathname.toLowerCase())
+  }
 
   if (redirect != null) {
     if (process.env.NODE_ENV !== `production`) {
@@ -62,7 +72,10 @@ const navigate = (to, options = {}) => {
   }
 
   let { pathname } = parsePath(to)
-  const redirect = redirectMap[pathname]
+  let redirect = redirectMap.get(pathname)
+  if (!redirect) {
+    redirect = redirectIgnoreCaseMap.get(pathname.toLowerCase())
+  }
 
   // If we're redirecting, just replace the passed in pathname
   // to the one we want to redirect to.