+++ packages/gatsby-admin/CHANGELOG.md
@@ -3,6 +3,10 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [0.10.0-next.9](https://github.com/gatsbyjs/gatsby/compare/gatsby-admin@0.10.0-next.8...gatsby-admin@0.10.0-next.9) (2021-02-25)
+
+**Note:** Version bump only for package gatsby-admin
+
 # [0.10.0-next.8](https://github.com/gatsbyjs/gatsby/compare/gatsby-admin@0.10.0-next.7...gatsby-admin@0.10.0-next.8) (2021-02-22)
 
 **Note:** Version bump only for package gatsby-admin